requirements.txt
appdirs==1.4.4
asgiref==3.5.1
asttokens==2.0.5
audioread==2.1.9
backcall==0.2.0
backports.zoneinfo==0.2.1
CacheControl==0.12.11
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
Cython==0.29.30
decorator==5.1.1
Django==4.0.4
executing==0.8.3
filelock==3.7.1
fonttools==4.33.3
h5py==3.7.0
hdmedians==0.14.2
huggingface-hub==0.7.0
idna==3.3
ipython==8.3.0
itermplot==0.5
jedi==0.18.1
joblib==1.1.0
kiwisolver==1.4.2
librosa==0.9.1
llvmlite==0.38.0
lockfile==0.12.2
lxml==4.8.0
matplotlib==3.5.2
matplotlib-inline==0.1.3
msgpack==1.0.4
natsort==8.1.0
nltk==3.7
numba==0.55.1
numpy==1.21.6
packaging==21.3
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pooch==1.6.0
praat-textgrids==1.3.1
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
resampy==0.2.2
scikit-learn==1.1.0
scipy==1.8.0
scitrack==2021.5.3
six==1.16.0
SoundFile==0.10.3.post1
sqlparse==0.4.2
stack-data==0.2.0
threadpoolctl==3.1.0
tinydb==4.7.0
tokenizers==0.12.1
torch==1.11.0
tqdm==4.64.0
traitlets==5.2.1.post0
transformers==4.19.2
typing_extensions==4.2.0
urllib3==1.26.9
uWSGI==2.0.20
wcwidth==0.2.5
